LightGCN/requirements.txt
pandas
scikit-learn
wandb
torch==1.12.1
torch_geometric
hydra-core